requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==5.12.0
gunicorn==21.2.0